Fix the habits column that rejected every new habit, and stop hiding refusals
The server was refusing the write. `clarity_habits.sort_order` was an
integer, while a new habit is stamped with Date.now() as its sort order
— about 1.79e12, far outside int4. Every habit created since has come
back 400 and stayed on whichever device made it. The edge logs show it
plainly: GETs 200, POSTs 400, over and over from the phone.

`clarity_tasks.sort_order` was already bigint, which is why tasks have
always synced and only habits went missing. The column is now bigint to
match, applied to the database and corrected in schema.sql.

The reason this took days to find is the more interesting half. Every
write ended in a swallowed catch, and the task and list writers did not
even look at the result — a row the server refused was indistinguishable
from one it accepted, so the caller marked it synced and never tried
again. Meanwhile Settings said "Tasks sync automatically across your
devices" unconditionally, which was simply untrue.

So writes now report. A refusal is recorded with the server's own words
and rethrown, so the caller requeues the row, and Settings shows "Not
syncing — changes are saved on this device only" with the message and a
retry. Reads report too. Had this existed it would have read "saving a
habit: value out of range for type integer" from the first attempt.

+++ supabase/schema.sql
@@ -102,7 +102,10 @@ create table if not exists public.clarity_habits (
   logs            jsonb not null default '{}'::jsonb,
   -- date -> short notes describing that day's logs.
   log_notes       jsonb not null default '{}'::jsonb,
-  sort_order      integer not null default 0,
+  -- bigint, not integer: a new habit's sort order is Date.now(), which is far
+  -- past the range of an int4. As an integer this rejected every habit ever
+  -- created with a 400, leaving it stranded on the device that made it.
+  sort_order      bigint not null default 0,
   -- 'notes' habits tick themselves when a note is written that day.
   source          text not null default 'manual'
                     check (source in ('manual', 'notes')),
